chore: split cache.go
